+++ tests/branching_kernels/if_else_2_labels/if_else_2_labels_dcmpl-gfx1010.cl
@@ -5,14 +5,14 @@ void if_else_2_labels(int x, __global int *data, int y)
     uint var3;
     uint var4;
     var3 = get_global_id(0);
-    if ((uint)1 != (uint)var3) {
+    if (1 != var3) {
         var4 = get_global_id(1);
         var0 = y * x;
     }
     else {
         var4 = get_global_id(1);
         var3 = 1;
-        var0 = (uint)(get_global_id(1) * x) - (uint)y;
+        var0 = (get_global_id(1) * x) - y;
     }
     data[var3] = var0;
     data[var4] = x;
