+++ tests/branching_kernels/if_else_and_if_else_0_labels/if_else_and_if_else_0_labels_hands-gfx.cl
@@ -6,16 +6,16 @@ void if_else_and_if_else_0_labels(int x, __global int *data, int y)
     int var5;
     uint var6;
     var3 = get_global_id(0);
-    if ((uint)1 != (uint)var3) {
+    if (1 != var3) {
         var5 = get_global_id(0) * y;
         var1 = get_global_id(1);
     }
     else {
         var1 = get_global_id(1);
         var3 = 1;
-        var5 = (uint)(get_global_id(1) * x) - (uint)y;
+        var5 = (get_global_id(1) * x) - y;
     }
-    var6 = (int)x < (int)y ? (ulong)x + (ulong)get_global_id(2) : (ulong)y + (ulong)var1;
+    var6 = x < y ? x + get_global_id(2) : y + var1;
     data[var3] = var5;
     data[get_global_id(2)] = var6;
     data[var1] = x;
